% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switchNet.R
\name{findChangePoints}
\alias{findChangePoints}
\title{Regime switching function to estimate the change points}
\usage{
findChangePoints(
  X,
  tau0,
  eps = c(5, 5),
  lambda = NULL,
  method = c("newton", "search"),
  maxIter = 30,
  both = TRUE,
  grid.tau1 = NULL,
  grid.tau2 = NULL
)
}
\arguments{
\item{X}{A p by N data matrix}

\item{tau0}{initialization of tau}

\item{eps}{nbhd size of approximation}

\item{lambda}{The tuning parameters in a vector of length 2}

\item{method}{The optimization method used. Could be "Newton" or "search",}

\item{maxIter}{The maximum number of iterations}

\item{grid.tau1}{}

\item{grid.tau2}{the search grid for tau1 and tau2. Must be provided if not using NR optimization}

\item{search}{Whether to perform grid search after Newton's update}
}
\description{
Regime switching function to estimate the change points
}
