% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switchNet.R
\name{switchNetBIC}
\alias{switchNetBIC}
\title{Compute the bic.score over a grid of lambda_1 and lambda_2}
\usage{
switchNetBIC(X, tau, lambda1, lambda2, tol = 1e-06)
}
\arguments{
\item{X}{A p by N data matrix}

\item{tau}{A length 2 vector of change points}

\item{lambda1}{A candidate sequence of tuning parameters for regime 1}

\item{lambda2}{A candidate sequence of tuning parameters for regime 2}

\item{tol}{Tolerance for declaring an edge}
}
\description{
Compute the bic.score over a grid of lambda_1 and lambda_2
}
\details{
This function
}
